requirements.txt
qrcode[pil]
Pillow
pyzbar
numpy
